Comment section and reccomendation system complete

+++ backend/movies/views.py
@@ -1,10 +1,13 @@
 import json
-
+import numpy as np 
+import pandas as pd
 
 from django.utils.formats import get_format
+from django.db.models import Case, When
 from datetime import datetime
 from rest_framework import status
-from rest_framework.decorators import api_view, permission_classes
+from rest_framework.decorators import api_view, permission_classes, authentication_classes
+from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 
@@ -13,6 +16,7 @@ from .models import *
 from .serializers import *
 
 from backend.app import IncomingApiManager as ApiM
+from .recom import *
 
 @api_view(('GET',))
 @permission_classes((AllowAny,))
@@ -100,6 +104,8 @@ def get_movie_details(request,movie_id):
 def get_max_results(**kwargs):
     if(kwargs['max_results'] is None):
         return 1
+    elif kwargs['max_results'] == None:
+    	return 1
     else:
         return int(kwargs['max_results'])
 
@@ -114,6 +120,7 @@ def get_all_comments(request):
 
 @api_view(('PUT',))
 @permission_classes((AllowAny,))
+@authentication_classes((TokenAuthentication, ))
 def edit_comment(request, comment_id):
 	"""
 	Incoming args: movie id
@@ -123,8 +130,14 @@ def edit_comment(request, comment_id):
 	except Comment.DoesNotExist:
 		return Response(status = status.HTTP_404_NOT_FOUND)
 
+	if request.user != comment.user:
+		return Response({'details': 'You do not have the permission'}, status=status.HTTP_401_UNAUTHORIZED)
+
 	if request.method == "PUT":
-		serializer = CommentSerializer(comment, data = request.data)
+		tempData = request.data.copy()
+		tempData['user'] = comment.user.pk
+		tempData['movie'] = comment.movie.pk
+		serializer = CommentSerializer(comment, data = tempData)
 		data = {}
 		if(serializer.is_valid()):
 			serializer.save()
@@ -135,20 +148,24 @@ def edit_comment(request, comment_id):
 
 @api_view(('DELETE',))
 @permission_classes((AllowAny,))
+@authentication_classes((TokenAuthentication, ))
 def delete_comment(request, comment_id):
 	"""
 	Incoming args: comment_id
 	"""
 	try:
-		comment = Comment.objects.filter(pk = comment_id)
+		comment = Comment.objects.get(pk = comment_id)
 	except Comment.DoesNotExist:
 		return Response(status = status.HTTP_404_NOT_FOUND)
+
+	if request.user != comment.user:
+			return Response({'details': 'You do not have the permission'}, status=status.HTTP_401_UNAUTHORIZED)
 
 	if request.method == "DELETE":
 		stat = comment.delete()
 		data = {}
 		if(stat):
-			data['success'] = 'update successful'
+			data['success'] = 'delete successful'
 		else:
 			data['failure'] = 'delete failed'
 		return Response(data = data)
@@ -156,22 +173,22 @@ def delete_comment(request, comment_id):
 
 @api_view(('POST','GET',))
 @permission_classes((AllowAny,))
+@authentication_classes((TokenAuthentication, ))
 def get_or_post_comment(request, movie_id):
 	"""
 	Info for generating movie instance should come from the frontend if request is POST
 	That is, [tmdb_id]
 	"""
+	try:
+		movie = Movie.objects.get(tmdb_id = movie_id)
+		print(movie)
+	except Movie.DoesNotExist:
+		data = ApiM.get_movie_details(movie_id)
+		movie = Movie.objects.create(tmdb_id = movie_id)
+		#call a view to fill the movie data. TODO://
 	if request.method == 'POST':
 		data = request.data
-		try:
-			movie = Movie.objects.get(tmdb_id = movie_id)
-			print(movie)
-		except Movie.DoesNotExist:
-			movie = Movie.objects.create(tmdb_id = movie_id)
-			movie.save()
-			#call a view to fill the movie data. TODO://
-
-		comment = Comment.objects.create(content = data['content'], movie = movie)
+		comment = Comment.objects.create(content = data['content'], movie = movie, user = request.user)
 		if comment:
 			return Response({'success':'successfully posted'}, status = status.HTTP_201_CREATED)
 		else:
@@ -181,10 +198,9 @@ def get_or_post_comment(request, movie_id):
 		Incoming args: movie id
 		"""
 		try:
-			comments = Comment.objects.filter(movie__tmdb_id = movie_id)
+			comments = Comment.objects.filter(movie = movie.pk)
 		except Comment.DoesNotExist:
 			return Response(status = status.HTTP_404_NOT_FOUND)
-
 		serializer = CommentSerializer(comments, many = True)
 		return Response(serializer.data)
 
@@ -193,15 +209,15 @@ def get_or_post_comment(request, movie_id):
 @permission_classes((AllowAny,))
 def get_comment_by_id(request, comment_id):
 	try:
-		comments = Comment.objects.get(pk = comment_id)
+		comment = Comment.objects.get(pk = comment_id)
 	except Comment.DoesNotExist:
 		return Response(status = status.HTTP_404_NOT_FOUND)
 
-	serializer = CommentSerializer(comments)
+	serializer = CommentSerializer(comment)
 	return Response(serializer.data)
 
 
-@api_view(('PUT',))
+@api_view(('POST',))
 @permission_classes((AllowAny,))
 def upvote_comment(request, comment_id):
 	try:
@@ -210,6 +226,7 @@ def upvote_comment(request, comment_id):
 		return Response(status = status.HTTP_404_NOT_FOUND)
 
 	if comment:
+		#Change contribution of comment owner. TODO://
 		comment.score += 1;
 		comment.save()
 		return Response({'success':'successfully upvoted'})
@@ -226,6 +243,7 @@ def downvote_comment(request, comment_id):
 		return Response(status = status.HTTP_404_NOT_FOUND)
 
 	if comment:
+		#Change contribution of comment owner. TODO://
 		comment.score -= 1;
 		comment.save()
 		return Response({'success':'successfully downvoted'})
@@ -243,3 +261,27 @@ def search_response_with_tmdb_id(request):
 
 	data = ApiM.search_response_with_tmdb_id(**kwargs)
 	return Response(data)
+
+# for recommendation
+@api_view(('GET',))
+@permission_classes((AllowAny,))
+def recommend(request):
+	df=pd.DataFrame(list(Rating.objects.all().values()))
+	nu=df.user_id.unique().shape[0]
+	current_user_id= request.user.id
+	# if new user not rated any movie
+	if current_user_id>nu:
+		movie=Movie.objects.get(id=1)
+		q=Rating(user=request.user,movie=movie,rating=0)
+		q.save()
+
+	print("Current user id: ",current_user_id)
+	prediction_matrix,Ymean = Myrecommend()
+	my_predictions = prediction_matrix[:,current_user_id-1]+Ymean.flatten()
+	pred_idxs_sorted = np.argsort(my_predictions)
+	pred_idxs_sorted[:] = pred_idxs_sorted[::-1]
+	pred_idxs_sorted=pred_idxs_sorted+1
+	preserved = Case(*[When(pk=pk, then=pos) for pos, pk in enumerate(pred_idxs_sorted)])
+	movie_list=list(Movie.objects.filter(id__in = pred_idxs_sorted,).order_by(preserved)[:10])
+	serializer = MoviesSerializer(movie_list, many = True)
+	return Response(serializer.data)
